+++ core/src/main/java/io/questdb/network/IODispatcherOsx.java
@@ -41,7 +41,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
     private final Kqueue kqueue;
     // the final ids are shifted by 1 bit which is reserved to distinguish
     // socket operations (0) and suspend events (1)
-    private long idSeq = 1;
+    private int idSeq = 1;
 
     public IODispatcherOsx(
             IODispatcherConfiguration configuration,
@@ -90,7 +90,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
         keventWriter.prepare();
         for (int i = watermark, sz = pending.size(); i < sz; i++) {
             final C context = pending.get(i);
-            final long id = pending.get(i, OPM_ID);
+            final int id = (int) pending.get(i, OPM_ID);
             final int operation = initialBias == IODispatcherConfiguration.BIAS_READ ? IOOperation.READ : IOOperation.WRITE;
             pending.set(i, OPM_OPERATION, operation);
 
@@ -106,7 +106,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
         keventWriter.done();
     }
 
-    private boolean handleSocketOperation(long id) {
+    private boolean handleSocketOperation(int id) {
         // find row in pending for two reasons:
         // 1. find payload
         // 2. remove row from pending, remaining rows will be timed out
@@ -167,7 +167,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
             return;
         }
 
-        final long opId = pendingEvents.get(eventsRow, EVM_OPERATION_ID);
+        final int opId = (int) pendingEvents.get(eventsRow, EVM_OPERATION_ID);
         final int row = pending.binarySearch(opId, OPM_ID);
         if (row < 0) {
             LOG.critical().$("internal error: suspended operation not found [id=").$(opId).$(", eventId=").$(id).I$();
@@ -190,11 +190,11 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
         pendingEvents.deleteRow(eventsRow);
     }
 
-    private long nextEventId() {
+    private int nextEventId() {
         return (idSeq++ << 1) + 1;
     }
 
-    private long nextOpId() {
+    private int nextOpId() {
         return idSeq++ << 1;
     }
 
@@ -241,7 +241,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
                 if (eventRow < 0) {
                     LOG.critical().$("internal error: suspend event not found on heartbeat [id=").$(opId).I$();
                 } else {
-                    final long eventId = pendingEvents.get(eventRow, EVM_ID);
+                    final int eventId = (int) pendingEvents.get(eventRow, EVM_ID);
                     keventWriter.prepare().readFD(suspendEvent.getFd(), eventId).done();
                     pendingEvents.deleteRow(eventRow);
                 }
@@ -271,7 +271,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
             interestSubSeq.done(cursor);
 
             useful = true;
-            long opId = nextOpId();
+            int opId = nextOpId();
             final long fd = context.getFd();
 
             int operation = requestedOperation;
@@ -323,7 +323,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
                 // if the operation was suspended, we request a read to be able to detect a client disconnect
                 operation = IOOperation.READ;
                 // ok, the operation was suspended, so we need to track the suspend event
-                final long eventId = nextEventId();
+                final int eventId = nextEventId();
                 LOG.debug().$("registering suspend event [fd=").$(fd)
                         .$(", op=").$(operation)
                         .$(", eventId=").$(eventId)
@@ -362,7 +362,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
     private void processSuspendEventDeadlines(long timestamp) {
         int count = 0;
         for (int i = 0, n = pendingEvents.size(); i < n && pendingEvents.get(i, EVM_DEADLINE) < timestamp; i++, count++) {
-            final long opId = pendingEvents.get(i, EVM_OPERATION_ID);
+            final int opId = (int) pendingEvents.get(i, EVM_OPERATION_ID);
             final int pendingRow = pending.binarySearch(opId, OPM_ID);
             if (pendingRow < 0) {
                 LOG.critical().$("internal error: failed to find operation for expired suspend event [id=").$(opId).I$();
@@ -384,7 +384,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
         pendingEvents.zapTop(count);
     }
 
-    private void rearmKqueue(C context, long id, int operation) {
+    private void rearmKqueue(C context, int id, int operation) {
         keventWriter.prepare();
 
         // Important: We must register for writing *before* registering for reading
@@ -427,7 +427,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
                 kqueue.setReadOffset(offset);
                 offset += KqueueAccessor.SIZEOF_KEVENT;
                 final long fd = kqueue.getFd();
-                final long id = kqueue.getData();
+                final int id = kqueue.getData();
                 // this is server socket, accept if there aren't too many already
                 if (fd == serverFd) {
                     accept(timestamp);
@@ -503,7 +503,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
             return lastError;
         }
 
-        public KeventWriter readFD(long fd, long id) {
+        public KeventWriter readFD(long fd, int id) {
             kqueue.setWriteOffset(offset);
             kqueue.readFD(fd, id);
             offset += KqueueAccessor.SIZEOF_KEVENT;
@@ -541,7 +541,7 @@ public class IODispatcherOsx<C extends IOContext<C>> extends AbstractIODispatche
             return this;
         }
 
-        public KeventWriter writeFD(long fd, long id) {
+        public KeventWriter writeFD(long fd, int id) {
             kqueue.setWriteOffset(offset);
             kqueue.writeFD(fd, id);
             offset += KqueueAccessor.SIZEOF_KEVENT;
